+++ Changes.md
@@ -1,4 +1,4 @@
-## 0.3.0
+## 0.3.0 - 2024-12-26
 
 - `ubi` now works with GitLab in addition to GitHub. For the command-line program, if the value you
   pass to `--project` has a `gitlab.com` domain, like `https://gitlab.com/gitlab-org/cli`, this will